Bump to 64 but containers. Major redo for CV and Geotiff container to avoid data loss when implicit downcasting double to float (GDAL default was set for 32bit).

+++ include/lad_processing.hpp
@@ -41,7 +41,7 @@ namespace lad
 
     KPlane computeFittingPlane (std::vector<KPoint> points);
 
-    std::vector<KPoint> convertMatrix2Vector (cv::Mat *matrix, float sx, float sy);
+    std::vector<KPoint> convertMatrix2Vector (cv::Mat *matrix, double sx, double sy);
 
     // Own implementation for scale independant landability analysis
